removed unused images api.py: moved plot code to features feature/features.py: has plot code
