--- data/sample_meta.json
@@ -1,12 +1,0 @@
-{
-    "Username": "KennyT",
-    "Latitude": "40.01",
-    "Longitude": "32.85",
-	"Altitude": "0.0",
-	"Year": 2021,
-	"Month": 3,
-	"Day": 20,
-	"Hour": 13,
-	"Minute": 24,
-	"Second": 46
-}
